+++ tests/mono/testcases/AddFloatsTest.cs
@@ -8,18 +8,18 @@ namespace TestCases {
 
         [Test()]
         public void test_null() {
-            Assert.AreEqual(service.add_floats(null, null), 0);
+            Assert.AreEqual(0, service.add_floats(null, null));
         }
 
         [Test()]
         public void test_empty() {
-            Assert.AreEqual(service.add_floats(0, 0), 0);
+            Assert.AreEqual(0, service.add_floats(0, 0));
         }
 
         [Test()]
         public void test() {
             Assert.AreEqual(
-                service.add_floats((float)10.5, (float)5.3), (float)15.8);
+                (float)15.8, service.add_floats((float)10.5, (float)5.3));
         }
     }
 }