+++ tests/mono/testcases/AddIntegerDictsTest.cs
@@ -9,16 +9,16 @@ namespace TestCases {
         [Test()]
         public void test_null_args() {
             ABIntegerDict r = service.add_integer_dicts(null, null);
-            Assert.AreEqual(r.a, 0);
-            Assert.AreEqual(r.b, 0);
+            Assert.AreEqual(0, r.a);
+            Assert.AreEqual(0, r.b);
         }
 
         [Test()]
         public void test_notset() {
             ABIntegerDict r = service.add_integer_dicts(
                 new ABIntegerDict(), new ABIntegerDict());
-            Assert.AreEqual(r.a, 0);
-            Assert.AreEqual(r.b, 0);
+            Assert.AreEqual(0, r.a);
+            Assert.AreEqual(0, r.b);
         }
 
         [Test()]
@@ -30,8 +30,8 @@ namespace TestCases {
             q.a = null;
             q.b = null;
             ABIntegerDict r = service.add_integer_dicts(p, q);
-            Assert.AreEqual(r.a, 0);
-            Assert.AreEqual(r.b, 0);
+            Assert.AreEqual(0, r.a);
+            Assert.AreEqual(0, r.b);
         }
 
         [Test()]
@@ -43,8 +43,8 @@ namespace TestCases {
             q.a = 0;
             q.b = 0;
             ABIntegerDict r = service.add_integer_dicts(p, q);
-            Assert.AreEqual(r.a, 0);
-            Assert.AreEqual(r.b, 0);
+            Assert.AreEqual(0, r.a);
+            Assert.AreEqual(0, r.b);
         }
 
         [Test()]
@@ -56,8 +56,8 @@ namespace TestCases {
             q.a = 50;
             q.b = 25;
             ABIntegerDict r = service.add_integer_dicts(p, q);
-            Assert.AreEqual(r.a, 150);
-            Assert.AreEqual(r.b, 75);
+            Assert.AreEqual(150, r.a);
+            Assert.AreEqual(75, r.b);
         }
     }
 }