+++ tests/mono/testcases/AddStringListsTest.cs
@@ -9,7 +9,7 @@ namespace TestCases {
         [Test()]
         public void test_null_args() {
             string[] r = service.add_string_lists(null, null);
-            Assert.AreEqual(r.Length, 0);
+            Assert.AreEqual(0, r.Length);
         }
 
         [Test()]
@@ -17,7 +17,7 @@ namespace TestCases {
             string[] a1 = {};
             string[] a2 = {};
             string[] r = service.add_string_lists(a1, a2);
-            Assert.AreEqual(r.Length, 0);
+            Assert.AreEqual(0, r.Length);
         }
 
         [Test()]
@@ -27,7 +27,7 @@ namespace TestCases {
             string[] r = service.add_string_lists(a1, a2);
             // That's weird... The client sends an empty array,
             // hence it receives an empty array as well
-            Assert.AreEqual(r.Length, 0);
+            Assert.AreEqual(0, r.Length);
         }
 
         [Test()]
@@ -35,8 +35,8 @@ namespace TestCases {
             string[] a1 = {""};
             string[] a2 = {""};
             string[] r = service.add_string_lists(a1, a2);
-            Assert.AreEqual(r.Length, 1);
-            Assert.AreEqual(r[0], "");
+            Assert.AreEqual(1, r.Length);
+            Assert.AreEqual("", r[0]);
         }
 
         [Test()]
@@ -45,9 +45,9 @@ namespace TestCases {
             string[] a2 = {"d", "e", "f"};
             string[] r = service.add_string_lists(a1, a2);
             Assert.AreEqual(r.Length, 3);
-            Assert.AreEqual(r[0], "ad");
-            Assert.AreEqual(r[1], "be");
-            Assert.AreEqual(r[2], "cf");
+            Assert.AreEqual("ad", r[0]);
+            Assert.AreEqual("be", r[1]);
+            Assert.AreEqual("cf", r[2]);
         }
 
         [Test()]
@@ -56,11 +56,11 @@ namespace TestCases {
             string[] a2 = {"d", "e", "f", "g", "h"};
             string[] r = service.add_string_lists(a1, a2);
             Assert.AreEqual(r.Length, 5);
-            Assert.AreEqual(r[0], "ad");
-            Assert.AreEqual(r[1], "be");
-            Assert.AreEqual(r[2], "cf");
-            Assert.AreEqual(r[3], "g");
-            Assert.AreEqual(r[4], "h");
+            Assert.AreEqual("ad", r[0]);
+            Assert.AreEqual("be", r[1]);
+            Assert.AreEqual("cf", r[2]);
+            Assert.AreEqual("g", r[3]);
+            Assert.AreEqual("h", r[4]);
         }
     }
 }