+++ tests/python/suds/testcases/add_string_lists.py
@@ -9,9 +9,7 @@ class AddStringListsTestCase(BaseTestCaseMixin, unittest.TestCase):
         p.item = []
         q = self.factory.create('types:StringList')
         q.item = []
-        # This one does not work, unfortunatelly,
-        # suds sends neither first argument nor second
-        #res = self.service.add_string_lists(p, q)
+        res = self.service.add_string_lists(p, q)
 
     def test(self):
         p = self.factory.create('types:StringList')